Add activity to handle view holder updater launch
Upon successful update of the device management role holder,
RESULT_OK is returned, otherwise the result is RESULT_CANCELED.

If the device management role holder updater is not currently
available (for example, if it is getting updated at this time),
there is retry logic in place. If none of the retries works,
RESULT_CANCELED is returned.

Test: atest RoleHolderUpdaterViewModelTest
Test: atest ViewModelEventTest
Bug: 204319866
Change-Id: I2c537baa41259e7e4edfb5a9668f7acfdb7efba0