Add required flag to registerReceiver call in DownloadPackageTask
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
unprotected broadcasts. This commit adds the RECEIVER_EXPORTED
flag to the call to registerReceiver in DownloadPackageTask since this
action is broadcast from the downloads provider.

Bug: 161145287
Test: Build
Change-Id: I304113cb96ef480a22017e834ed82dfbb22a0a0a