Always start role holder after a role holder-triggered update
Today we only start the role holder a second time if it requested
an update the first time and the update succeeded. We should instead
always start the role holder a second time after it requested an
update, regardless of whether the update was successful or not.
This way the role holder can determine whether it should perform
provisioning via its older version, or fail altogether.

Fixes: 224993192
Test: manual
Change-Id: I11043ca8912e449184f7a2e688e048317a412039