Avoiding DevicePolicyManager.resetDefaultCrossProfileIntentFilters for clone profile during OTA.
Observed that DevicePolicyManager.resetDefaultCrossProfileIntentFilters
is called for clone profile, based on investigation observed that this
was supposed to be invoked from owner profile. This is called from OTA
controller for managed profile, where for any user which is not managed
profile we call DevicePolicyManager.resetDefaultCrossProfileIntentFilters, the intention was to update default CrossProfileIntentFilters between
managed profile and onwer. Changing the flow to only allow
DevicePolicyManager.resetDefaultCrossProfileIntentFilters being called
for users that are not profile.

DevicePolicyManager.resetDefaultCrossProfileIntentFilters is called for any profile, it would created self-loop i.e instead of configuring filters between profile and owner it also configures filters between profile and itself leading to no cross profile intent resolution. This was happening for clone profile, where there were no filters configured between owner and clone but only filters configured between clone and itself.

Bug: 267630556
Test: Manually tested by setting property persist.pm.mock-upgrade and
flasing the device.

Change-Id: Ic5294eda2ad2c06519d19ac830cb98380ca0d87c