Fix Voltron on HSUM builds.
Currently DO provisioning via Voltron is not working on HSUM builds if
DPC app is not bundled, due to the fact that the Voltron flow assumes
the system user being an initiator of the provisioning flow.
On HSUM builds the initiator is not a system user, but the first
active user user10.

Conduct the flow on the current active user instead of the system user.

Test: manual provisioning
Flag: EXEMPT bugfix
Bug: b/374256023

Change-Id: I6dedcd02f6b3ec1502a81b092d3f18cb3ec32413